--- a/package.json
+++ b/package.json
@@ -38,6 +38,7 @@
     "strip-bom": "^3.0.0",
     "tar": "^2.2.1",
     "tar-stream": "^1.5.2",
+    "v8-compile-cache": "^1.0.0",
     "validate-npm-package-license": "^3.0.1"
   },
   "devDependencies": {