**Summary**

This change makes it so that `yarn` can take advantage of V8's compile cache, and improve startup times. From https://raw.githubusercontent.com/zertosh/v8-compile-cache/v1.0.0/README.md:

> `v8-compile-cache` attaches a `require` hook to use [V8's code cache](https://v8project.blogspot.com/2015/07/code-caching.html) to speed up instantiation time. The "code cache" is the work of parsing and compiling done by V8.
> The ability to tap into V8 to produce/consume this cache was introduced in [Node v5.7.0](https://nodejs.org/en/blog/release/v5.7.0/).

`time`time ./bin/yarn.js config list` goes from 0.22s to 0.18s. config list` goes from 0.22s to 0.18s.

**Test plan**

`./bin/yarn.js`, `./bin/yarn.js add underscore` and `./bin/yarn.js config list`. All work.